pyproject.toml
[build-system]
requires = ["setuptools>=64.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "perception-for-planning"
version = "0.1.0"
description = "Perception tools for robotic planning using Gemini Vision API"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "NishanthJKumar"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    # Google GenAI SDK for Gemini API
    "google-genai>=0.6.0",
    
    # Image processing
    "Pillow>=9.0.0",
    
    # HTTP requests and utilities
    "requests>=2.25.0",
    
    # Numerical and scientific computing
    "numpy>=1.21.0",
    "scipy>=1.6.0",
    
    # Visualization
    "matplotlib>=3.5.0",
    "supervision>=0.17.0",  # For segmentation visualization
    
    # 3D processing
    "open3d>=0.16.0",       # Point cloud processing
    "trimesh>=3.9.0",       # 3D mesh handling
    
    # API integrations
    "replicate>=0.18.0",    # For SAM-2 via Replicate API
    
    # Segment Anything Model
    "segment-anything>=1.0",
]

[project.urls]
"Bug Reports" = "https://github.com/NishanthJKumar/perception_for_planning/issues"
"Source" = "https://github.com/NishanthJKumar/perception_for_planning"
"Documentation" = "https://ai.google.dev/gemini-api/docs/robotics-overview"

[tool.setuptools]
packages = ["perception_for_planning"]

[tool.setuptools.package-data]
perception_for_planning = ["py.typed"]

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py311"]

[tool.isort]
profile = "black"
line_length = 100